+++ leviculum-std/src/lib.rs
@@ -32,6 +32,7 @@ pub(crate) mod ini_config;
 pub mod interfaces;
 pub(crate) mod known_destinations;
 pub(crate) mod packet_hashlist;
+pub mod process;
 pub mod remote_status;
 pub mod resource_policy;
 pub mod reticulum;